fix(pose_estimation/generators/pose_estimation/tfr_map_methods): type of cutout image

+++ pose_estimation/generators/pose_estimation/tfr_map_methods.py
@@ -904,7 +904,7 @@ class DropBlockPostMethod(TFRPostMapMethod):
         # `where` in this case faster than simple multiplication, because we must cast boolean array
         cutout_image = tf.where(
             add_color_dims_boolean,
-            tf.zeros_like(image, dtype=tf.uint8),
+            tf.zeros_like(image, dtype=image.dtype),
             image
         )
